Add operator checker for frozen graph file
Add a checker to check whether the operators in the
frozen graph file are supported in the current
frontend handlers. The input can be either binary
(.pb, .ckpt) or text (.pbtxt). All unsupported operators
will be shown in the console output. It will also
verify the arguments for the supported operators.