Add auto_pad attribute support to Conv
Add auto_pad attribute support to Conv so it will be used along with
pads to determine the padding.

The assumption for ConvTranspose is that auto_pad has higher precedence
than output_shape and output_padding. That is when auto_pad is provided
and it is not NOTSET, the output shape will always be inferred.