Fix map_fn with string issue on gpu (#654)
The tfidf vectorizer handler has a map_fn with string constant
that doesn't work on gpu only. It is a known issue in Tensorflow.
This PR is to provide a temporary fix by placing the map_fn
specifically on cpu. Once the TF bug is fixed, we can remove
the fix.